% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general_data_wrangle.R
\name{archive}
\alias{archive}
\title{Archive Filter for Residents}
\usage{
archive(data)
}
\arguments{
\item{data}{A dataframe. The dataset containing the resident information.}
}
\value{
A character vector of archived resident names.
}
\description{
Filters out archived residents from a dataset. It assumes the 
dataset contains a 'res_archive' column with the value "Yes" for archived residents.
}
\examples{
\dontrun{
 data <- tibble::tibble(name = c("John", "Jane", "Jake"), res_archive = c("No", "Yes", "No"))
 archived_residents <- archive(data)
 print(archived_residents)
}
}
