% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mod_support.R
\name{calculate_resident_level}
\alias{calculate_resident_level}
\title{Calculate Resident Level}
\usage{
calculate_resident_level(coach_data)
}
\arguments{
\item{coach_data}{A dataset containing resident data, including `type` and `grad_yr`}
}
\value{
A dataset with an additional `Level` column
}
\description{
Determines the level of residents (Intern, PGY2, PGY3) based on their type and graduation year.
}
