% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general_data_wrangle.R
\name{fill_missing_resident_data}
\alias{fill_missing_resident_data}
\title{General Data Wrangling Utilities}
\usage{
fill_missing_resident_data(data, remove_incomplete = TRUE)
}
\arguments{
\item{data}{A dataframe containing the resident data.}

\item{remove_incomplete}{Logical. If TRUE, removes any rows that still have missing 'name' or 'record_id'.}
}
\value{
A dataframe with filled 'name' and 'record_id' columns.
}
\description{
This file contains utility functions for data cleaning prior to visualization and analysis.
}
\details{
## Function Overview:
- `fill_missing_resident_data`: Ensures each row in a dataset has a 'name' and 'record_id'.
- `merge_old_new_data`: Merges old data saved as an RDS file with new REDCap API data.
- `archive`: Filters out archived residents from datasets.

These utilities are essential for ensuring complete and accurate data for visualization and analysis.

Fill Missing Names and Record IDs

Ensures that each row in the provided dataframe has complete 'name' and 'record_id' columns.
The function fills missing 'name' and 'record_id' using lookup logic, where information 
is propagated within the dataframe using `left_join()` and `tidyr::fill()`.
}
\examples{
\dontrun{
 data <- tibble::tibble(record_id = c(1, 2, 3, NA), name = c(NA, "John Doe", NA, "Jane Smith"))
 filled_data <- fill_missing_resident_data(data)
 print(filled_data)
}
}
