% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_data_pulls.R
\name{forms_api_pull}
\alias{forms_api_pull}
\title{Forms API Pull}
\usage{
forms_api_pull(token, url, ..., fields = NULL)
}
\arguments{
\item{token}{A string. The API token for accessing the RedCap project.}

\item{url}{A string. The API URL for the RedCap project.}

\item{...}{Character names of forms to pull.}

\item{fields}{A character vector of field names to pull.}
}
\value{
A dataframe of the requested forms.
}
\description{
Pulls data for specific forms from RedCap.
}
