% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_data_pulls.R
\name{full_api_pull}
\alias{full_api_pull}
\title{Full API Pull}
\usage{
full_api_pull(token, url)
}
\arguments{
\item{token}{A string. The API token for accessing the RedCap project.}

\item{url}{A string. The API URL for the RedCap project.}
}
\value{
A dataframe of the full data.
}
\description{
Pulls all data from RedCap using the provided token and API URL.
}
