% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_submit.R
\name{generate_new_instance}
\alias{generate_new_instance}
\title{Generate New Instance}
\usage{
generate_new_instance(
  record_id,
  instrument_name,
  coach_data,
  redcap_uri,
  token
)
}
\arguments{
\item{record_id}{A numeric. The REDCap record ID.}

\item{redcap_uri}{A string. The REDCap API URL.}

\item{token}{A string. The REDCap API token.}

\item{instrument}{A string. The REDCap instrument name.}
}
\value{
The next available instance for the repeating instrument as an integer.
}
\description{
Calculate the next available instance for a specific instrument and record.
}
