% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_data_pulls.R
\name{generate_survey_link}
\alias{generate_survey_link}
\title{Generate Survey Link}
\usage{
generate_survey_link(record_id, instrument, url, token)
}
\arguments{
\item{record_id}{A string. The record ID.}

\item{instrument}{A string. The instrument name.}

\item{url}{A string. The API URL for the RedCap project.}

\item{token}{A string. The API token for accessing the RedCap project.}
}
\value{
A list containing the survey link.
}
\description{
Generates a new survey link for a given record and instrument.
}
