% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_milestone_wrangle.R
\name{get_all_milestones}
\alias{get_all_milestones}
\title{Get All Milestones}
\usage{
get_all_milestones(token, url)
}
\value{
A data frame with all milestone data.
}
\description{
Pulls data from REDCap for the specified milestone forms.
}
