% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_data_pulls.R
\name{get_links_api}
\alias{get_links_api}
\title{Get Links API}
\usage{
get_links_api(token, url, record_id, instrument)
}
\arguments{
\item{token}{A string. The API token for accessing the RedCap project.}

\item{url}{A string. The API URL for the RedCap project.}

\item{record_id}{A string. The record ID.}

\item{instrument}{A string. The instrument name.}
}
\value{
A survey link.
}
\description{
Pulls survey links from RedCap.
}
