% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_milestone.R
\name{get_milestone_columns}
\alias{get_milestone_columns}
\title{Get Milestone Columns}
\usage{
get_milestone_columns(data, type = "program")
}
\arguments{
\item{data}{The data frame containing milestone data.}

\item{type}{The type of milestone, either "program" or "self".}
}
\value{
A vector of milestone column names.
}
\description{
Extracts the milestone columns for either "program" or "self" milestone data.
}
