% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{imres_colors}
\alias{imres_colors}
\title{imres Color Palette}
\format{
An object of class \code{list} of length 7.
}
\usage{
imres_colors
}
\description{
Standard colors used in imres applications
}
\keyword{datasets}
