% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{imres_css}
\alias{imres_css}
\title{Load Custom imres CSS}
\usage{
imres_css()
}
\value{
A tags$link object pointing to the CSS file
}
\description{
This function loads the custom CSS file for imres applications.
}
