% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{imres_logo}
\alias{imres_logo}
\title{Get Path to imres Logo}
\usage{
imres_logo()
}
\value{
A character string with the path to the logo file
}
\description{
This function returns the file path to the imres logo.
}
