% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{imres_theme}
\alias{imres_theme}
\title{Apply Standard imres Theme}
\usage{
imres_theme()
}
\value{
A tags$style object with HTML CSS
}
\description{
This function applies the standard imres styling to a Shiny application.
}
