% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general_data_wrangle.R
\name{merge_old_new_data}
\alias{merge_old_new_data}
\title{Merge Old and New Data}
\usage{
merge_old_new_data(
  old_data_path,
  token,
  url,
  timestamp_columns,
  merge_key = "record_id"
)
}
\arguments{
\item{old_data_path}{A string. The file path to the `.rds` file containing the old data.}

\item{token}{A string. The API token for accessing the REDCap project.}

\item{url}{A string. The API URL for the REDCap project.}

\item{timestamp_columns}{A character vector. The names of timestamp columns used to check for updates.}

\item{merge_key}{A string. The key to use for merging old and new data (default is "record_id").}
}
\value{
A data frame containing the merged data, with duplicates removed.
}
\description{
Merges old data (saved as an `.rds` file) with new data pulled from the REDCap API. 
It performs a full join on a user-defined key (default is "record_id") and 
removes any duplicate rows from the combined dataset.
}
\examples{
\dontrun{
 merged_data <- merge_old_new_data(
   old_data_path = "data/old_data.rds", 
   token = "your_redcap_token", 
   url = "https://redcap.yourinstitution.edu/api/", 
   timestamp_columns = c("event_timestamp", "another_timestamp")
 )
}
}
