% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_milestone.R
\name{miles_plot}
\alias{miles_plot}
\title{Milestones Radar Plot}
\usage{
miles_plot(data, name, period)
}
\arguments{
\item{data}{A data frame containing the milestone data.}

\item{name}{The resident name for which to generate the radar plot.}

\item{period}{The milestone period to filter the data.}
}
\value{
A ggplot object representing the radar plot.
}
\description{
This function generates a radar plot using the `ggradar` package for a given
resident and period. The input data is filtered internally using `name` and `period`.
}
