% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_milestone.entry.R
\name{mod_miles_rating_server}
\alias{mod_miles_rating_server}
\title{Module Server for Milestone Rating (simplified version)}
\usage{
mod_miles_rating_server(id, period)
}
\arguments{
\item{id}{module id}

\item{period}{reactive returning the selected period}
}
\description{
Module Server for Milestone Rating (simplified version)
}
