% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_milestone.entry.R
\name{mod_miles_rating_ui}
\alias{mod_miles_rating_ui}
\title{Module UI for Milestone Rating (simplified version)}
\usage{
mod_miles_rating_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Module UI for Milestone Rating (simplified version)
}
