% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_submit.R
\name{prepare_data_for_submission}
\alias{prepare_data_for_submission}
\title{Prepare Data for Submission}
\usage{
prepare_data_for_submission(data, record_id)
}
\arguments{
\item{data}{A data frame. The input data to be formatted for submission.}

\item{record_id}{A character or numeric. The REDCap record ID.}
}
\value{
A list with formatted data ready for REDCap submission.
}
\description{
Format the input data for submission to REDCap.
}
