% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_assessment.R
\name{process_cc_document_data}
\alias{process_cc_document_data}
\title{Process Continuity Clinic Documentation Data}
\usage{
process_cc_document_data(data, resident_name)
}
\arguments{
\item{data}{A data frame containing evaluation data.}

\item{resident_name}{A character string specifying the resident's name.}
}
\value{
A data frame with documentation data including Level, Update EMR (renamed from cc_doc_update),
Review EMR (renamed from cc_doc_review), Update meds (renamed from cc_doc_medall), Does reminders (renamed from cc_doc_remind),
Accurate Diagnoses (renamed from cc_doc_diag), Notes clear (renamed from cc_doc_notes), Notes complete (renamed from cc_doc_comp),
and Evaluator.
}
\description{
Filters and processes documentation evaluation data for the continuity clinic for a given resident.
}
