% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_assessment.R
\name{process_cc_inbasket_data}
\alias{process_cc_inbasket_data}
\title{Process Continuity Clinic Inbasket Data}
\usage{
process_cc_inbasket_data(data, resident_name)
}
\arguments{
\item{data}{A data frame containing evaluation data.}

\item{resident_name}{A character string specifying the resident's name.}
}
\value{
A data frame with inbasket data including Level, Responsible (renamed from cc_inb_resp),
Result Response (renamed from cc_inb_resu), MyChart (renamed from cc_inb_mych), Paperwork (renamed from cc_inb_comm),
and Comms with preceptor (renamed from cc_inb_comp).
}
\description{
Filters and processes inbasket evaluation data for the continuity clinic for a given resident.
}
