% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_milestone_wrangle.R
\name{process_milestones}
\alias{process_milestones}
\title{Process Milestone Data}
\usage{
process_milestones(data, type)
}
\arguments{
\item{data}{The data frame containing milestone data.}

\item{type}{The type of milestone, "program" or "self".}
}
\value{
A processed milestone dataset.
}
\description{
Processes "program" or "self" milestone data, appending medians and removing archived residents.
}
