% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_data_pulls.R
\name{redcap_api_call}
\alias{redcap_api_call}
\title{RedCap API Utilities}
\usage{
redcap_api_call(token, url, formData)
}
\arguments{
\item{token}{A string. The API token for accessing the RedCap project.}

\item{url}{A string. The API URL for the RedCap project.}

\item{formData}{A list. The form data to send in the POST request.}
}
\value{
A dataframe of the response data.
}
\description{
This file contains utility functions for interacting with the RedCap API.

Makes a generic POST request to RedCap API with provided form data.
}
\details{
Function Overview:
- `redcap_api_call()`: Handles the generic API call to RedCap.
- `full_api_pull()`: Pulls all data from RedCap.
- `forms_api_pull()`: Pulls data for specific forms from RedCap.
- `fetch_new_redcap_data()`: Incremental data pull from RedCap based on timestamps.
- `get_links_api()`: Gets survey links for a specific instrument and record.
- `generate_survey_link()`: Generates a new survey link for a specific record and instrument.
}
