% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general_data_wrangle.R
\name{remove_archived_residents}
\alias{remove_archived_residents}
\title{Remove Archived Residents}
\usage{
remove_archived_residents(data)
}
\arguments{
\item{data}{A dataframe. The dataset containing the resident information.}
}
\value{
A dataframe with archived residents removed.
}
\description{
Removes rows of archived residents from a dataset, based on the `archive` function.
}
\examples{
\dontrun{
 data <- tibble::tibble(name = c("John", "Jane", "Jake"), res_archive = c("No", "Yes", "No"))
 data_no_archives <- remove_archived_residents(data)
 print(data_no_archives)
}
}
