% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_milestone.R
\name{rename_milestone_columns}
\alias{rename_milestone_columns}
\title{Rename Milestone Columns}
\usage{
rename_milestone_columns(df)
}
\arguments{
\item{df}{A data frame with milestone columns.}
}
\value{
A data frame with renamed columns.
}
\description{
Renames milestone columns to clean up names for plotting and analysis.
}
