% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_submit.R
\name{submit_to_redcap}
\alias{submit_to_redcap}
\title{Submit Data to REDCap}
\usage{
submit_to_redcap(data, record_id, redcap_uri, token)
}
\arguments{
\item{data}{A data frame. The data to submit to REDCap.}

\item{record_id}{A numeric. The record ID.}

\item{redcap_uri}{A string. The REDCap API URL.}

\item{token}{A string. The REDCap API token.}
}
\value{
A response object containing the status of the submission.
}
\description{
Submit a data frame to REDCap.
}
