% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_submit.R
\name{update_inputs}
\alias{update_inputs}
\title{Update Inputs}
\usage{
update_inputs(session, data, input_fields)
}
\arguments{
\item{session}{The Shiny session object.}

\item{data}{A data frame. The existing REDCap data.}

\item{input_fields}{A character vector of input field names.}
}
\description{
Update Shiny UI inputs with existing REDCap data.
}
