% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_redcap_submit.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Inputs}
\usage{
validate_inputs(inputs, valid_range = 0:9)
}
\arguments{
\item{inputs}{A named list of input values.}

\item{valid_range}{A vector. The acceptable range of input values.}
}
\value{
A list with `success` (TRUE/FALSE) and invalid inputs.
}
\description{
Validate that inputs are within the required range.
}
