% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_assessment.R
\name{create_styled_dt}
\alias{create_styled_dt}
\title{Create a Styled DataTable for CC Evaluation Data}
\usage{
create_styled_dt(data, caption = NULL)
}
\arguments{
\item{data}{A data frame of processed evaluation data.}

\item{caption}{An optional character string for the table caption/title.}
}
\value{
A DT::datatable object with applied styling.
}
\description{
Generates a DT::datatable object with custom styling and an optional caption.
}
