% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_assessment.R
\name{generate_p_d}
\alias{generate_p_d}
\title{Generate Plus-Delta Data}
\usage{
generate_p_d(data, resident)
}
\arguments{
\item{data}{A data frame containing evaluation data.}

\item{resident}{A character string specifying the resident's name.}
}
\value{
A data frame containing the Date, Rotation, Level, Plus (renamed from cc_res_does_well),
Delta (renamed from res_to_improve), Feedback (renamed from min_giv_feedback), and Evaluator for the resident.
}
\description{
Filters and processes the input data to extract plus-delta evaluations for a specific resident.
}
